Feat[#2] : 관측소명을 Query로 보내도록 수정했습니다.
