+++ Kloudy/Kloudy/Extension/FetchWeatherInformation.swift
@@ -10,8 +10,8 @@ import Foundation
 class FetchWeatherInformation {
     func startLoad(city: String) {
         // 도시 이름을 받아서 x, y값 받음
-        let xyCoordinate = findCityCoordinate(city: city)
-        let xCoordinate = xyCoordinate[0], yCoordinate = xyCoordinate[1]
+        let cityInformation = getCityInformaiton(city: city)
+        let xCoordinate = cityInformation[0], yCoordinate = cityInformation[1], airConditionMeasuring = cityInformation[2]
         let dayTime = getNowTimeForQuery()
         let day = dayTime[0], time = dayTime[1]
         
@@ -20,11 +20,13 @@ class FetchWeatherInformation {
         
         let xQuery = URLQueryItem(name: "x", value: xCoordinate)
         let yQuery = URLQueryItem(name: "y", value: yCoordinate)
+        let airQuery = URLQueryItem(name: "air", value: airConditionMeasuring)
         let dayQuery = URLQueryItem(name: "day", value: day)
         let timeQuery = URLQueryItem(name: "time", value: time)
         
         urlComponents?.queryItems?.append(xQuery)
         urlComponents?.queryItems?.append(yQuery)
+        urlComponents?.queryItems?.append(airQuery)
         urlComponents?.queryItems?.append(dayQuery)
         urlComponents?.queryItems?.append(timeQuery)
         
@@ -58,17 +60,12 @@ class FetchWeatherInformation {
         dataTask.resume()
     }
     
-    func findCityCoordinate(city: String) -> [String] {
-        // 나중에 엑셀파일에서 긁어오겠음
-        let koreaCitySamples: [CityInformation] = [
-            CityInformation(name: "서울시", x: 60, y: 127),
-            CityInformation(name: "인천시", x: 55, y: 124),
-            CityInformation(name: "수원시", x: 60, y: 121)
-        ]
+    func getCityInformaiton(city: String) -> [String] {
+        let koreanCities: [CityInformation] = loadCityListFromCSV()
         
-        let nowCity = koreaCitySamples.filter{ $0.name == city }[0]
+        let nowCity = koreanCities.filter{ $0.city == city }[0]
         
-        return [String(nowCity.x), String(nowCity.y)]
+        return [String(nowCity.xCoordination), String(nowCity.yCoordination), nowCity.airCoditionMeasuring]
     }
     
     func getNowTimeForQuery() -> [String] {
@@ -98,5 +95,32 @@ class FetchWeatherInformation {
         }
         
         return result
+    }
+    
+    private func parseCSVAt(url: URL) -> [CityInformation] {
+        var cityList: [CityInformation] = []
+        
+        do {
+            // url을 받은 data
+            let data = try Data(contentsOf: url)
+            // 해당 data를 encoding 합니다.
+            let dataEncoded = String(data: data, encoding: .utf8)
+            if let dataArr = dataEncoded?.components(separatedBy: "\n").map({$0.components(separatedBy: ",")}) {
+                                for item in dataArr {
+                                    let city = CityInformation(code: item[0], province: item[1], city: item[2], airCoditionMeasuring: item[3], xCoordination: Int(item[4]) ?? 0, yCoordination: Int(item[5]) ?? 0, longitude: Double(item[6]) ?? 0, latitude: Double(item[7]) ?? 0)
+                                    cityList.append(city)
+                                }
+                            }
+        } catch {
+            print("Error reading CSV file")
+        }
+        
+        return cityList
+    }
+    
+    private func loadCityListFromCSV() -> [CityInformation] {
+        // bundle에 있는 경로 > Calorie라는 이름을 가진 csv 파일 경로
+        let path = Bundle.main.path(forResource: "CityInformation", ofType: "csv")!
+        return parseCSVAt(url: URL(fileURLWithPath: path))
     }
 }
